pyproject.toml
[project]
name = "deepseek-ocr"
version = "0.1.0"
description = "FastAPI web application for OCR using deepseek-ocr via langchain-ollama"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "langchain>=0.1.0",
    "langchain-ollama>=0.1.0",
    "pillow>=10.0.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

